requirements.txt
paho-mqtt
playsound
PyObjC
gin
